ui/scripts/requirements.txt
rembg>=2.0.62
Pillow>=10.0.0
